refactor(mongo): turn transformDateExtract into a factory method
to overwrite some part when used with mongo 5
(next commit)
